net: keystone: pa: changes to support timestamping by multiple netcp modules
In addition to PA, other netcp modules such as CPTS or PHY can be capable of
timestamping the packets.

In the tx direction, if a module such as CPTS or PHY closer to the wire claims
timestamping of a packet, it marks the packet by a SKBTX_IN_PROGRESS flag and
PA will not timestamp that packet.

In the rx direction, a module closer to the wire marks the packet in the
packet info's rxtstamp_complete field after timestamping a packet and PA
will not timestamp that packet.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>