tci6638: dts: Reconfigure accumulators for greater NETCP efficiency
With the pending switch to TX completion interrupts in NETCP, a behavioral
change in the accumulators is needed to maintain efficiency. The two
changes implemented by this patch are to permit a delay of up to about 1ms
before generating an interrupt, and an increase in the buffer size from 8
entries to 16. These numbers can be tuned as the end-user sees fit.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>