net: keystone: Use TX completion interrupts to reap completed packets
Prior to this patch, TX completion queues were reaped only when the
pool of free descriptors dropped below dma_poll_threshold (normally
1/2 of the free pool size). This could result in lengthy delays before
completing the associated SKBs, which is not acceptable when reporting
CPTS-generated timestamps. Further, this resulted in irregular ICMP
ECHO REPLY generation starting in the 3.4 kernel.

This patch removes the use of dma_poll_threshold and uses accumulator
queue interrupts to trigger reaping of completed packets. The accumulator
configuration can be changed in the DTS to tune this behavior, although
QPEND queues could be used for the highest-priority queues to maximize
throughput.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>