net: keystone: Optimize psdata alignment function
This patch adds special cases for the most common alignments
to allow the compiler to generate more-efficient shift operators
rather than a slower divide/modulo calculation. Since this occurs
on every packet transmitted through the PA (i.e. ALL of them), this
optimization is worth the possible reduction in obviousness.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>