net: keystone: cpsw: add support of CPTS timestamping
The Common Platform Time Sync (CPTS) module is used to facilitate
host control of time sync operations.  It enable compliance with
the IEEE 1588-2008 standard for a precision clock synchronization
protocol.

This patch adds the support of timestamping packets in the tx and
rx data path by the CPTS module that match the PTP packet filter
requirement.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>