net: ethernet: cpts: clean up event list if event pool is empty
When a CPTS user does not exit gracefully by disabling cpts timestamping
and leaving a joined multicast group, the system continues to receive
and timestamps the ptp packets which eventually occupy all the event
list entries.  When this happns, the added code tries to remove some
list entries which are expired.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>