ARM: bL_switcher: Allow detachment of outstanding switch requests
In rare situations, it may be neecssary to prevent a completer
registered using bL_switch_request_cb() from being called.

For example, a cpufreq driver module needs to ensure that its
completer won't get called if the module is about to be unloaded
from the kernel.

This patch provides a bL_switch_request_detach() function which
detaches the completer from an outstanding request.  The request
continues in the background, as if it had been started using
bL_switch_request().

If somebody else's completer is registered instead then the caller
shouldn't care about it, so nothing is done in this case.

Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Nicolas Pitre <nicolas.pitre@linaro.org>