tci6638: dts: Fix "dma dma0chan4: out of descriptors" error
The NETCP PA configuration code does not have a graceful means to handle
PktDMA descriptor exhaustion. According to the documentation it's called
with a spinlock held so it can't sleep, and allocating a large number of
descriptors for a rare configuration activity is wasteful.

To ensure rapid return of free descriptors, this patch assigns the
"patx-pdsp0" channel's completion queue to a qpend queue.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>