net: keystone: Fix issue with dma flags
when sending an inflow IPSec packet to SA, protocol specific flags in the descriptor should never be set. SA engine treats this as an error condition, moreover these flags are not carried over when the packet leaves SA. Current version of the driver was setting this flag to indicate the egress CPSW port to switch. But in reality this had no effect whatsover on the packet routing and it would cause the SA to drop the packet if this was ever set to a non-zero value.
To send directed packets to CPSW after SA processing, the packet must be sent to PDSP-5 with appropriate routing command to set the PSFLAGs. This is already being done by the ethernet driver, hence the above logic was redundant.