remoteproc: add mmap support for rproc memory
The remoreproc-user used to expose the device memory
thru UIO interface. But, the UIO mmap interface does
not provide option to configure start address of the
mapped region. The change in rproc-user provides mmap
interface to user for the configured rproc memory regions.

One of the usecase is remotecore elf loader from user
space. In non-iommu case the CMA buffers can't be
used as the loader will need to load elf segments to
start address pre-determined by the segment address of
the image.

Signed-off-by: Sajesh Kumar Saran <sajesh@ti.com>
Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>