pandaboard: secondary_startup endian fix
If kernel operates in BE mode on board that has LE bootloader/rom code, we need
to switch CPU to operate in BE mode as soon as possible. secondary_startup
that is called from omap specific secondary startup code will do the switch,
but we need it to do earlier because OMAP specific secondary_startup
code works with BE data.

Signed-off-by: Victor Kamensky <victor.kamensky@linaro.org>