net: keystone: Allow MTU changes when interface is down
The netcp_ndo_change_mtu() function asserted a BUG if the interface
was down when the MTU was changed. This has never been necessary,
but neither has time been available to fix it. This patch fixes it.

This patch also replaces the magic number 9500 with the proper define
for the maximum size of a network frame the CPSW can process.

Signed-off-by: Reece R. Pollack <x0183204@ti.com>