initramfs: remove unnecessary symlinks processing shortcut
Special handling of symlinks in do_header() assumes that name and body
entries are sequential and reads them together. This shortcut has no
real performance benefits, but it complicates changes to the state
machine.

Make handling of symlinks more similar to a regular files. Store name
in name_buf and destination in symlink_buf.

Change-Id: I9ef93fa73b098a475fbb306d0b4ad0cb934ad38c
Signed-off-by: Taras Kondratiuk <takondra@cisco.com>